Add tests for parsing state
